+++ noodles-fasta/src/repository/adapters.rs
@@ -1,5 +1,6 @@
 //! Sequence repository adapters.
 
+mod empty;
 mod indexed_reader;
 
-pub use self::indexed_reader::IndexedReader;
+pub use self::{empty::Empty, indexed_reader::IndexedReader};